Add docker password

+++ scripts/travis-build.sh
@@ -3,8 +3,11 @@ set -e
 
 build_docker() {
     docker build -t $DOCKER_REPO:$1 --build-arg device=$1 app
-    docker push $DOCKER_REPO:$1
-    docker rmi -f $DOCKER_REPO:$1
+
+    if [ "$TRAVIS_BRANCH" = "master" -a "$TRAVIS_PULL_REQUEST" = "false" ]; then
+        docker push $DOCKER_REPO:$1
+        docker rmi -f $DOCKER_REPO:$1
+    fi
 }
 
 docker login -u="$DOCKER_USER" -p="$DOCKER_PASS"
